Scaffold website with NextJS starter code
I'm admittedly no expert on frontend frameworks, but my brief googling
found that this one can build us a static site and seems to have wide
adoption so it hopefully wouldn't be jarring to future contributors with
more frontend experience. I considered trying to just handroll building
the HTML/CSS/JS since this is truly just a single static page, but the
value of having a mature build system seemed worthwhile and will help
with integrating things like CodeMirror for the editor windows.