Proof-of-concept integrating formatting into the sandbox
This is very much a rough proof-of-concept -- putting the formatting
behind a manual button press is not optimal but it demonstrates that it
can work in this package. As we restructure this UI to look more like an
actual code editor, making auto-formatting an option with some
debouncing on user input delay is probably the correct long-term
solution.